+++ src/main/java/minhdo/swe/project/repository/SubMemberRepository.java
@@ -1,25 +1,24 @@
 package minhdo.swe.project.repository;
 
 import minhdo.swe.project.entity.Sub;
-import minhdo.swe.project.entity.SubMember;
+import minhdo.swe.project.entity.SubMembership;
 import minhdo.swe.project.entity.User;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
 import java.util.Optional;
 
-public interface SubMemberRepository extends JpaRepository<SubMember, Long> {
+public interface SubMemberRepository extends JpaRepository<SubMembership, Long> {
     long countBySub(Sub sub);
 
     boolean existsByUserAndSub(User user, Sub sub);
 
-    boolean existsByUserAndSubAndRole(User user, Sub sub, SubMember.Role role);
+    boolean existsByUserAndSubAndRole(User user, Sub sub, SubMembership.Role role);
 
-    Optional<SubMember> findByUserAndSub(User user, Sub sub);
+    Optional<SubMembership> findByUserAndSub(User user, Sub sub);
 
     @EntityGraph(attributePaths = {"user"})
-    Page<SubMember> findAllBySub(Sub sub, Pageable pageable);
+    Page<SubMembership> findAllBySub(Sub sub, Pageable pageable);
 }