+++ src/main/java/minhdo/swe/project/service/SubMembershipService.java
@@ -1,0 +1,4 @@
+package minhdo.swe.project.service;
+
+public class SubMembershipService {
+}